LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-mouse-operations.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/common.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=date id=date value="2000-01-02">
<script>
description("Tests if calendar picker key bindings work as expected.");

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

openPicker(document.getElementById('date'), test);

function test() {
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '2000-01-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-02');

    debug('Check that hovering over an entry highlights it.');
    hoverOverDayCellAt(5, 0);
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '1999-12-31');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-31');

    hoverOverDayCellAt(2, 1);
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '2000-01-04');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-04');

    debug('Check that navigating between months maintains highlighted position.');
    eventSender.keyDown('m');
    shouldBeEqualToString('currentMonth()', '2000-02');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2000-02-08');
    shouldBeEqualToString('highlightedDayCells()', '2000-02-08');

    debug('Check that mouse click sets the value.');
    clickDayCellAt(6, 0);
    shouldBeEqualToString('currentMonth()', '2000-02');
    shouldBeEqualToString('selectedValue()', '2000-02-05');
    shouldBeEqualToString('selectedDayCells()', '2000-02-05');
    shouldBeEqualToString('highlightedValue()', '2000-02-05');
    shouldBeEqualToString('highlightedDayCells()', '2000-02-05');
    shouldBeEqualToString('document.getElementById("date").value', '2000-02-05');

    finishJSTest();
}
</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
