LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-mouse-operations-expected.txt

Tests if calendar picker key bindings work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "2000-01-02"
PASS highlightedDayCells() is "2000-01-02"
Check that hovering over an entry highlights it.
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "1999-12-31"
PASS highlightedDayCells() is "1999-12-31"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "2000-01-04"
PASS highlightedDayCells() is "2000-01-04"
Check that navigating between months maintains highlighted position.
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2000-02-08"
PASS highlightedDayCells() is "2000-02-08"
Check that mouse click sets the value.
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-02-05"
PASS selectedDayCells() is "2000-02-05"
PASS highlightedValue() is "2000-02-05"
PASS highlightedDayCells() is "2000-02-05"
PASS document.getElementById("date").value is "2000-02-05"
PASS successfullyParsed is true

TEST COMPLETE
 
